Move breaadcrumb swtich block and return block html element to a new Breadcrumb.js component. Import and call it in Page, Form & Show components.
