+++ client/src/components/Customers/Form.js
@@ -16,6 +16,8 @@ import { CountryDropdown, RegionDropdown } from 'react-country-region-selector'
 import $ from 'jquery'
 $.fn.checkbox = require('semantic-ui-checkbox')
 
+import Breadcrumb from '../Layouts/Breadcrumb'
+
 class Form extends Component {
   constructor(props) {
     super(props)
@@ -183,7 +185,10 @@ class Form extends Component {
     //    { key: 'delivered', value: 'delivered', text: 'DELIVERED' } ]
 
     return (  
-      <div className="row">
+      <div className="ui stackable grid">
+
+        <Breadcrumb />
+
         <div className="ui text container ui segment">  
 
           <form className={classnames("ui form", { loading: isLoading })} onSubmit={this.handleSubmit.bind(this)}>
