+++ client/src/components/Customers/Page.js
@@ -7,6 +7,8 @@ import { fetchCustomers } from '../../actions/customerActions'
 // Localization 
 import T from 'i18n-react'
 
+import Breadcrumb from '../Layouts/Breadcrumb'
+
 // Datatables
 import $ from 'jquery'
 window.JSZip = require('jszip')
@@ -29,7 +31,7 @@ class Page extends Component {
     
     this.props.fetchCustomers()
 
-    $('.table').DataTable({
+    $('.table').dataTable({
       processing: true,
       responsive: true,
       language: {
@@ -86,6 +88,9 @@ class Page extends Component {
     
     return (
       <div className="row column">  
+
+        <Breadcrumb />
+
         <div className="ui vertical segment">
           <Link className="ui primary button" to="/customers/new">
             <i className="add circle icon"></i>
