+++ client/src/components/Projects/Form.js
@@ -14,6 +14,8 @@ import 'react-datepicker/dist/react-datepicker.css'
 // Localization 
 import T from 'i18n-react'
 
+import Breadcrumb from '../Layouts/Breadcrumb'
+
 class Form extends Component {
   constructor(props) {
     super(props)
@@ -105,7 +107,10 @@ class Form extends Component {
     )
     
     return (
-      <div className="row">
+      <div className="ui stackable grid">
+
+        <Breadcrumb />
+
         <div className="ui text container ui segment">  
 
           <form className={classnames("ui form", { loading: isLoading })} onSubmit={this.handleSubmit.bind(this)}>
